backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
boto3==1.33.1
langgraph==0.0.24
langchain==0.1.7
langchain-google-genai==0.0.13
google-generativeai==0.3.1
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
nltk==3.8.1
python-multipart==0.0.6
uvicorn==0.24.0
aiofiles==23.2.1
requests==2.31.0
